Add inst. for containers & unordered-containers
This means instances for:
* Map, HashMap, IntMap
* Set, HashSet, IntSet

The definition of those types is the same for the lazy and the strict
one, so they all only have one instance. I use foldl everywhere since
it is not possible to import the internal module.

I only added tests for Set and Map since the code for the other
containers is almost exactly the same.

This commit also includes minor cleanups in the cabal file and the
Data.LargeHashable.Class module.